requirements.txt
streamlit>=1.28.0
tensorflow>=2.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
pillow>=10.0.0
pandas>=2.0.0
plotly>=5.0.0
# Note: TensorFlow works with Python 3.8-3.12
# For Python 3.12+, use: python3.12 -m venv venv

